:recycle: Separate sign and encrypt modules
