+++ src/crypto/index.js
@@ -13,8 +13,9 @@
  *
  */
 import * as convert from './convert';
+import * as encrypt from './encrypt';
 import hash from './hash';
 import * as keys from './keys';
 import * as sign from './sign';
 
-export default Object.assign({}, convert, { hash }, keys, sign);
+export default Object.assign({}, convert, encrypt, { hash }, keys, sign);
