+++ test/crypto/sign.js
@@ -19,17 +19,11 @@ import {
 	verifyMessageWithTwoPublicKeys,
 	printSignedMessage,
 	signAndPrintMessage,
-	encryptMessageWithPassphrase,
-	decryptMessageWithPassphrase,
 	signData,
 	verifyData,
-	encryptPassphraseWithPassword,
-	decryptPassphraseWithPassword,
 } from '../../src/crypto/sign';
 
-const convert = require('../../src/crypto/convert');
 const keys = require('../../src/crypto/keys');
-const hash = require('../../src/crypto/hash');
 
 const makeInvalid = str => {
 	const char = str[0] === '0' ? '1' : '0';
@@ -82,14 +76,11 @@ ${defaultSecondSignature}
 	const defaultData = Buffer.from('This is some data');
 	const defaultDataSignature =
 		'b8704e11c4d9fad9960c7b6a69dcf48c1bede5b74ed8974cd005d9a407deef618dd800fe69ceed1fd52bb1e0881e71aec137c35b90eda9afe93716a5652ee009';
-	const defaultPassword = 'myTotal53cr3t%&';
 
 	let defaultSignedMessage;
 	let defaultDoubleSignedMessage;
-	let defaultEncryptedMessageWithNonce;
 
 	let getPrivateAndPublicKeyBytesFromPassphraseStub;
-	let hashStub;
 
 	beforeEach(() => {
 		defaultSignedMessage = {
@@ -104,27 +95,6 @@ ${defaultSecondSignature}
 			signature: defaultSignature,
 			secondSignature: defaultSecondSignature,
 		};
-		defaultEncryptedMessageWithNonce = {
-			encryptedMessage:
-				'299390b9cbb92fe6a43daece2ceaecbacd01c7c03cfdba51d693b5c0e2b65c634115',
-			nonce: 'df4c8b09e270d2cb3f7b3d53dfa8a6f3441ad3b14a13fb66',
-		};
-		sandbox
-			.stub(convert, 'convertPrivateKeyEd2Curve')
-			.returns(
-				Buffer.from(
-					'd8be8cacb03fb02f34e85030f902b635f364d6c23f090c7640e9dc9c568e7d5e',
-					'hex',
-				),
-			);
-		sandbox
-			.stub(convert, 'convertPublicKeyEd2Curve')
-			.returns(
-				Buffer.from(
-					'f245e78c83196d73452e55581ef924a1b792d352c142257aa3af13cded2e7905',
-					'hex',
-				),
-			);
 
 		getPrivateAndPublicKeyBytesFromPassphraseStub = sandbox.stub(
 			keys,
@@ -142,15 +112,6 @@ ${defaultSecondSignature}
 				privateKey: Buffer.from(defaultSecondPrivateKey, 'hex'),
 				publicKey: Buffer.from(defaultSecondPublicKey, 'hex'),
 			});
-
-		hashStub = sandbox
-			.stub(hash, 'default')
-			.returns(
-				Buffer.from(
-					'd43eed9049dd8f35106c720669a1148b2c6288d9ea517b936c33a1d84117a760',
-					'hex',
-				),
-			);
 	});
 
 	describe('#signMessageWithPassphrase', () => {
@@ -358,111 +319,6 @@ ${defaultSecondSignature}
 				defaultPublicKey,
 			);
 			verification.should.be.true();
-		});
-	});
-
-	describe('#encryptMessageWithPassphrase', () => {
-		let encryptedMessage;
-
-		beforeEach(() => {
-			encryptedMessage = encryptMessageWithPassphrase(
-				defaultMessage,
-				defaultPassphrase,
-				defaultPublicKey,
-			);
-		});
-
-		it('should encrypt a message', () => {
-			encryptedMessage.should.have
-				.property('encryptedMessage')
-				.be.hexString()
-				.with.length(68);
-		});
-
-		it('should output the nonce', () => {
-			encryptedMessage.should.have
-				.property('nonce')
-				.be.hexString()
-				.with.length(48);
-		});
-	});
-
-	describe('#decryptMessageWithPassphrase', () => {
-		it('should be able to decrypt the message correctly using the receiver’s secret passphrase', () => {
-			const decryptedMessage = decryptMessageWithPassphrase(
-				defaultEncryptedMessageWithNonce.encryptedMessage,
-				defaultEncryptedMessageWithNonce.nonce,
-				defaultPassphrase,
-				defaultPublicKey,
-			);
-
-			decryptedMessage.should.be.equal(defaultMessage);
-		});
-	});
-
-	describe('encrypt and decrypt passphrase with password', () => {
-		beforeEach(() => {
-			hashStub.returns(
-				Buffer.from(
-					'e09dfc943d65d63f4f31e444c81afc6d5cf442c988fb87180165dd7119d3ae61',
-					'hex',
-				),
-			);
-		});
-
-		describe('#encryptPassphraseWithPassword', () => {
-			let cipher;
-
-			beforeEach(() => {
-				cipher = encryptPassphraseWithPassword(
-					defaultPassphrase,
-					defaultPassword,
-				);
-			});
-
-			it('should encrypt a passphrase', () => {
-				cipher.should.be
-					.type('object')
-					.and.have.property('cipher')
-					.and.be.hexString();
-			});
-
-			it('should output the IV', () => {
-				cipher.should.be
-					.type('object')
-					.and.have.property('iv')
-					.and.be.hexString()
-					.and.have.length(32);
-			});
-		});
-
-		describe('#decryptPassphraseWithPassword', () => {
-			it('should decrypt a text with a password', () => {
-				const cipherAndNonce = {
-					cipher:
-						'1c527b9408e77ae79e2ceb1ad5907ec523cd957d30c6a08dc922686e62ed98271910ca5b605f95aec98c438b6214fa7e83e3689f3fba89bfcaee937b35a3d931640afe79c353499a500f14c35bd3fd08',
-					iv: '89d0fa0b955219a0e6239339fbb8239f',
-				};
-				const decrypted = decryptPassphraseWithPassword(
-					cipherAndNonce,
-					defaultPassword,
-				);
-				decrypted.should.be.eql(defaultPassphrase);
-			});
-		});
-
-		describe('integration test', () => {
-			it('should encrypt a given passphrase with a password and decrypt it back to the original passphrase', () => {
-				const encryptedString = encryptPassphraseWithPassword(
-					defaultPassphrase,
-					defaultPassword,
-				);
-				const decryptedString = decryptPassphraseWithPassword(
-					encryptedString,
-					defaultPassword,
-				);
-				decryptedString.should.be.eql(defaultPassphrase);
-			});
 		});
 	});
 });
